Added a speed modifier - and a previous track button
I made it so speed modifier only can reduce speed,
since if it was possible to increase speed with a slider it would be
hard to get it back to normal.
A potential solution would just be have set speed increments of x2, x4,
x8 etc, but I like the slider so im going to keep that.

Previous track button, just required a bit of extension on the existing
next track logic, through a bool, which determines if i want to
increment or decrement the track index and clamp respectively