Implemented Multiple Background Images into the database
Idea came about, when i was thinking, if you put a song in the system,
with lyrics, you could change the background to match the lyrics of the
song.

Currently the interval rate is set to perfectly divide between the
number of images you have to cycle through. Since I am using it merely
for a change of graphic.

Originally I was going to try to implement an Audio Visualiser, using
the frequency and amplitude of the sound waves, but that wasn't feasible
considerring the scope of this project