Okay, this one was confusing to implement - Fading
Implemented a complete fading system.
So depending on the stage of the current track you are on, it will fade
from previous image, to appropriate image, based on a bunch of index
maths.

This was causing me alot of grief, since, the fading, would blur into
itself, if you moved the timeline bar to a point, that required the same
image as the current one, furthermore, if you tried to jump from image
index 0, to image index 2, it would snap to image index 1 then fade to
2, which was  really odd.

Eventually figured it all out - no noticeable bugs remain. Need to tidy
this up though xD